Update LEDiva.ino
Select a Color:
Press and hold the Color button.
The LEDs will all turn on and become dim.
Press the Pattern button to cycle through the colors.
Release the Color button to lock in your selection.

Select a Pattern:
Press and hold the Pattern button.
Wait until the LEDs finish the current animation pattern and remain static.
Press the Color button to cycle through the patterns. Note: Each pattern will play through one cycle. Wait until the LEDs stop changing before advancing to the next patten.
Release the Pattern button to lock in your selection.

Change the max brightness of the LEDs:
Press and hold the Color button, then turn the LEDiva on.
The LEDs will turn white.
Keep holding the Color button to reduce the brightness of the LEDs.
Press the Pattern button to increase the brightness of the LEDs.
Once finished, release all buttons and wait for two seconds.
The LEDiva will save your selection and start up normally.

Change the number of LEDs used:
Press and hold the Pattern button, then turn the LEDiva on.
The LEDs will turn white.
Keep holding the Pattern button to reduce the number of LEDs
Press the Color button to increase the number of LEDs (up to 100).
Once finished, release all buttons and wait for two seconds.
The LEDiva will save your selection and start up normally.